+++ src/GafferModule/OptionalValuePlugBinding.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2013, Image Engine Design Inc. All rights reserved.
+//  Copyright (c) 2023, Cinesite VFX Ltd. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
@@ -34,35 +34,11 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "GafferSceneTest/TestShader.h"
+#pragma once
 
-#include "Gaffer/CompoundNumericPlug.h"
-#include "Gaffer/OptionalValuePlug.h"
-#include "Gaffer/StringPlug.h"
-#include "Gaffer/SplinePlug.h"
-
-using namespace IECore;
-using namespace Gaffer;
-using namespace GafferSceneTest;
-
-GAFFER_NODE_DEFINE_TYPE( TestShader )
-
-TestShader::TestShader( const std::string &name )
-	:	Shader( name )
+namespace GafferModule
 {
-	// The base class expects us to serialise a `loadShader()`
-	// call to set the values for these, but we just represent
-	// a fixed shader. Turn serialisation back on.
-	namePlug()->setFlags( Plug::Serialisable, true );
-	typePlug()->setFlags( Plug::Serialisable, true );
 
-	addChild( new Color3fPlug( "out", Plug::Out ) );
-	parametersPlug()->addChild( new IntPlug( "i" ) );
-	parametersPlug()->addChild( new Color3fPlug( "c" ) );
-	parametersPlug()->addChild( new SplinefColor3fPlug( "spline" ) );
-	parametersPlug()->addChild( new OptionalValuePlug( "optionalString", new Gaffer::StringPlug() ) );
-}
+void bindOptionalValuePlug();
 
-TestShader::~TestShader()
-{
-}
+} // namespace GafferModule