+++ Changes.md
@@ -15,6 +15,12 @@ Fixes
   - Fixed translation of `vector` typed outputs defined as `vector <name>` in an output definition.
   - Fixed translation of `shadow:enable` and `shadow:color` parameters on UsdLux lights, which were previously ignored.
 
+API
+---
+
+- OptionalValuePlug : Added a new plug type that pairs an `enabled` BoolPlug with a `value` ValuePlug.
+- Shader : Added support for using OptionalValuePlug to represent optional parameters.
+
 1.3.0.0 (relative to 1.2.10.0)
 =======